+++ library/src/main/java/org/mustangproject/TradeParty.java
@@ -12,6 +12,7 @@ import org.mustangproject.ZUGFeRD.IZUGFeRDTradeSettlement;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 /***
@@ -424,6 +425,7 @@ public class TradeParty implements IZUGFeRDExportableTradeParty {
 
 
 	@Override
+	@JsonIgnore
 	public String getUriUniversalCommunicationID() {
 		if (uriUniversalCommunicationId != null) {
 			return uriUniversalCommunicationId.getID();
@@ -433,6 +435,7 @@ public class TradeParty implements IZUGFeRDExportableTradeParty {
 
 
 	@Override
+	@JsonIgnore
 	public String getUriUniversalCommunicationIDScheme() {
 		if (uriUniversalCommunicationId != null) {
 			return uriUniversalCommunicationId.getScheme();